refactor: move MapObjectsButton to ui/buttons folder
